add class TableBase
TableBase is a base class based on QTableWidget.
It's used to display some objects in a very general/abstract way.
Each object is displayed as a row of the table where each column shows
a custom "view" of the object. The different views for the columns are
specified from outside. Delete/Update logic is also specified from
outside. The TableBase class just handles the UI of deleting/updating
and calls the corresponding functuons specified from outside.
Additionally it has support for different-sized rows.